+++ net.tourbook.chart/src/net/tourbook/chart/ChartMarker.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2005, 2011  Wolfgang Schramm and Contributors
+ * Copyright (C) 2005, 2012  Wolfgang Schramm and Contributors
  * 
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -20,7 +20,7 @@ public class ChartMarker {
 	/**
 	 * x-position in graph units
 	 */
-	public float	graphX;
+	public double	graphX;
 
 	/**
 	 * index in the data serie
