+++ net.tourbook.statistics/src/net/tourbook/statistics/StatisticWeekHrZone.java
@@ -156,10 +156,10 @@ public class StatisticWeekHrZone extends YearStatistic {
 //		};
 	}
 
-	private float[] createWeekData() {
+	private double[] createWeekData() {
 
 		final int weekCounter = _resortedHrZoneValues[0].length;
-		final float allWeeks[] = new float[weekCounter];
+		final double allWeeks[] = new double[weekCounter];
 
 		for (int weekIndex = 0; weekIndex < weekCounter; weekIndex++) {
 			allWeeks[weekIndex] = weekIndex;
